python-requirements.txt
nameko
redis
mysql-connector-python
pyjwt
bcrypt
celery==4.4.6